moved SVGs into components folder and renamed them in kebab-case
