+++ src/components/SVGs/big-green-curve.tsx
