+++ src/components/SVGs/drill.tsx
