+++ src/components/SVGs/gear.tsx
